+++ open_mastr/soap_api/config.py
@@ -89,21 +89,6 @@ def get_data_config():
     return data_config
 
 
-def get_db_tables():
-    """
-    Names of database tables
-    
-    Returns
-    -------
-    dict
-        Table names
-    """
-    with open(os.path.join(get_project_home_dir(), "config", "tables.yml")) as data_fh:
-        db_tables = yaml.safe_load(data_fh)
-
-    return db_tables
-
-
 def create_project_home_dir():
     """Create directory structure of PROJECTHOME
     """
@@ -126,7 +111,7 @@ def create_project_home_dir():
     log.info(f'I will create a default set of config files in {config_path}')
 
     internal_config_dir = os.path.join(pathlib.Path(__file__).parent.absolute(), 'config')
-    files = ["data.yml", "tables.yml", "logging.yml"]
+    files = ["data.yml", "logging.yml"]
 
     for file in files:
         if not file in os.listdir(config_path):
@@ -211,6 +196,12 @@ def _filenames_generator():
                     # Collect file names for all technologies
                     filenames[section].update({tech: tmp})
 
+        # Add file names of cleaned data
+        filenames["cleaned"] = {
+            tech: f"{prefix}_{tech}_cleaned.csv" for tech in get_power_unit_types()
+        }
+
+        # Add file names of processed data
         filenames["postprocessed"] = {
             tech: f"{prefix}_{tech}.csv" for tech in get_power_unit_types()
         }